+++ webapp/channels/src/components/more_direct_channels/list/list.tsx
@@ -2,12 +2,13 @@
 // See LICENSE.txt for license information.
 
 import _ from 'lodash';
+import type {ReactNode} from 'react';
 import React, {useCallback, useEffect, useMemo} from 'react';
 import {FormattedMessage, useIntl} from 'react-intl';
 
 import type {UserProfile} from '@mattermost/types/users';
 
-import MultiSelect from 'components/multiselect/multiselect';
+import MultiSelect from 'components/multiselect';
 
 import Constants from 'utils/constants';
 
@@ -17,11 +18,12 @@ import ListItem from '../list_item';
 import type {Option, OptionValue} from '../types';
 import {optionValue} from '../types';
 
-const MAX_SELECTABLE_VALUES = Constants.MAX_USERS_IN_GM - 1;
+const MAX_SELECTABLE_VALUES = Constants.MAX_USERS_IN_GM;
 export const USERS_PER_PAGE = 50;
 
 export type Props = {
     addValue: (value: OptionValue) => void;
+    children?: ReactNode;
     currentUserId: string;
     handleDelete: (values: OptionValue[]) => void;
     handlePageChange: (page: number, prevPage: number) => void;
@@ -118,15 +120,25 @@ const List = React.forwardRef((props: Props, ref?: React.Ref<MultiSelect<OptionV
             handleAdd={props.addValue}
             handleSubmit={props.handleSubmit}
             noteText={note}
+            disableMultiSelectList={props.values.length > (Constants.MAX_USERS_IN_GM - 1)}
             maxValues={MAX_SELECTABLE_VALUES}
+            changeMessageColor='red'
+            showError={props.values.length === MAX_SELECTABLE_VALUES}
             numRemainingText={
-                <FormattedMessage
-                    id='multiselect.numPeopleRemaining'
-                    defaultMessage='Use ↑↓ to browse, ↵ to select. You can add {num, number} more {num, plural, one {person} other {people}}. '
-                    values={{
-                        num: MAX_SELECTABLE_VALUES - props.values.length,
-                    }}
-                />
+                props.values.length === MAX_SELECTABLE_VALUES ? (
+                    <FormattedMessage
+                        id='multiselect.noMorePeople'
+                        defaultMessage='A personal message is limited to {maxUsers} people.'
+                        values={{
+                            maxUsers: (Constants.MAX_USERS_IN_GM - 1),
+                        }}
+                    />
+                ) : (
+                    <FormattedMessage
+                        id='multiselect.numPeopleRemaining'
+                        defaultMessage='Use the ↑↓ arrows on the keyboard, ENTER to select.'
+                    />
+                )
             }
             buttonSubmitText={
                 <FormattedMessage
@@ -146,7 +158,10 @@ const List = React.forwardRef((props: Props, ref?: React.Ref<MultiSelect<OptionV
             users={props.users}
             totalCount={props.totalCount}
             placeholderText={intl.formatMessage({id: 'multiselect.placeholder', defaultMessage: 'Search and add members'})}
-        />
+
+        >
+            {props.children}
+        </MultiSelect>
     );
 });