+++ webapp/channels/src/components/post_view/post_list/post_list.tsx
@@ -109,7 +109,7 @@ export interface Props {
     toggleShouldStartFromBottomWhenUnread: () => void;
     shouldStartFromBottomWhenUnread: boolean;
     hasInaccessiblePosts: boolean;
-
+    isThreadView?: boolean;
     actions: {
 
         /*
@@ -376,6 +376,7 @@ export default class PostList extends React.PureComponent<Props, State> {
                             latestPostTimeStamp={this.props.latestPostTimeStamp}
                             isMobileView={this.props.isMobileView}
                             lastViewedAt={this.props.lastViewedAt}
+                            isThreadView={this.props.isThreadView}
                         />
                     </div>
                 </div>