+++ webapp/channels/src/components/root/root.tsx
@@ -78,6 +78,7 @@ import RootRedirect from './root_redirect';
 import {checkIKTokenExpiresSoon, checkIKTokenIsExpired, clearLocalStorageToken, getChallengeAndRedirectToLogin, isDefaultAuthServer, refreshIKToken, storeTokenResponse} from '../login/utils';
 
 import 'plugins/export.js';
+import {LLMBotPost} from 'plugins/ai/components/llmbot_post';
 
 const LazyErrorPage = React.lazy(() => import('components/error_page'));
 const LazyLogin = React.lazy(() => import('components/login/login'));
@@ -754,8 +755,10 @@ export default class Root extends React.PureComponent<Props, State> {
         this.initiateMeRequests();
 
         // See figma design on issue https://mattermost.atlassian.net/browse/MM-43649
+
         // this.props.actions.registerCustomPostRenderer('custom_up_notification', OpenPricingModalPost, 'upgrade_post_message_renderer');
         // this.props.actions.registerCustomPostRenderer('custom_pl_notification', OpenPluginInstallPost, 'plugin_install_post_message_renderer');
+        this.props.actions.registerCustomPostRenderer('custom_llmbot', LLMBotPost, 'llmbot_post_message_renderer');
 
         if (this.themeMediaQuery.addEventListener) {
             this.themeMediaQuery.addEventListener('change', this.handleThemeMediaQueryChangeEvent);