+++ webapp/channels/src/packages/mattermost-redux/src/constants/websocket.ts
@@ -27,6 +27,8 @@ const WebsocketEvents = {
     ROLE_UPDATED: 'role_updated',
     TYPING: 'client-user_typing',
     STOP_TYPING: 'stop_typing',
+    RECORDING: 'client-user_recording',
+    STOP_RECORDING: 'stop_recording',
     PREFERENCE_CHANGED: 'preference_changed',
     PREFERENCES_CHANGED: 'preferences_changed',
     PREFERENCES_DELETED: 'preferences_deleted',