+++ webapp/platform/client/src/client4.ts
@@ -53,7 +53,7 @@ import {
     ServerChannel,
     PendingGuests,
 } from '@mattermost/types/channels';
-import {Options, StatusOK, ClientResponse, LogLevel, FetchPaginatedThreadOptions} from '@mattermost/types/client4';
+import {Options, StatusOK, ClientResponse, LogLevel, FetchPaginatedThreadOptions, SummarizeResult} from '@mattermost/types/client4';
 import {Compliance} from '@mattermost/types/compliance';
 import {
     ClientConfig,
@@ -2369,12 +2369,19 @@ export default class Client4 {
         );
     }
 
-    addPostReminder = (userId: string, postId: string, timestamp: number) => {
+    addPostReminder = (userId: string, postId: string, timestamp: number, reschedule?: boolean, reminderPostId?: string) => {
         this.trackEvent('api', 'api_post_set_reminder');
 
         return this.doFetch<StatusOK>(
             `${this.getUserRoute(userId)}/posts/${postId}/reminder`,
-            {method: 'post', body: JSON.stringify({target_time: timestamp})},
+            {method: 'post', body: JSON.stringify({target_time: timestamp, reschedule: reschedule, post_id: reminderPostId})},
+        );
+    }
+
+    markPostReminderAsDone = (userId: string, postId: string) => {
+        return this.doFetch<StatusOK>(
+            `${this.getUserRoute(userId)}/posts/${postId}/reminder`,
+            {method: 'delete'},
         );
     }
 
@@ -4857,6 +4864,33 @@ export default class Client4 {
                 client_ids: clientId,
             })}
         )
+    }
+
+    //
+    // PLUGIN AI
+    //
+    async doSummarize(postId: string, botUsername: string): Promise<SummarizeResult> {
+        const url = `${this.getPostRoute(postId)}/summarize?botUsername=${botUsername}`;
+
+        return this.doFetch(url, {method: 'post'});
+    }
+
+    async doStopGenerating(postId: string) {
+        const url = `${this.getPostRoute(postId)}/stop`;
+
+        return this.doFetch(url, {method: 'post'});
+    }
+
+    async doRegenerate(postId: string) {
+        const url = `${this.getPostRoute(postId)}/regenerate`;
+
+        return this.doFetch(url, {method: 'post'});
+    }
+
+    async doPostbackSummary(postId: string) {
+        const url = `${this.getPostRoute(postId)}/postback_summary`;
+
+        return this.doFetch(url, {method: 'post'});
     }
 }